tmpfs_size (size_t) + indent

+++ common.h
@@ -83,7 +83,7 @@ struct nsjconf_t {
 	uid_t initial_uid;
 	gid_t initial_gid;
 	unsigned int max_conns_per_ip;
-	unsigned int tmpfs_size;
+	size_t tmpfs_size;
 	 LIST_HEAD(pidslist, pids_t) pids;
 	 LIST_HEAD(rwbindmountptslist, constchar_t) rwbindmountpts;
 	 LIST_HEAD(robindmountptslist, constchar_t) robindmountpts;
